move merkle tree class to helper functions
