+++ block/block.py
@@ -2,7 +2,7 @@ import hashlib
 import sys
 import json
 import datetime as date
-import merkleTree
+from helpers import findMerkleRoot
 
 # define variables
 version = "00000001"  # version
@@ -45,7 +45,7 @@ class Block(object):
         self.timestamp = str(date.datetime.now()).encode('utf8')
 
     def merkle_root(self):
-        self.merkle_root = merkleTree.findMerkleRoot(self.transactions)
+        self.merkle_root = findMerkleRoot(self.transactions)
 
     def block_version(self, vers):
         # gets the version and updates that instance with the version
